Incorporating review comments/ changes
